+++ Nodes/ControllerNode.js
@@ -18,6 +18,7 @@ module.exports = function(Polyglot) {
   const Vehicle = require('./Vehicle.js')(Polyglot);
   const VehicleSecurity = require('./VehicleSecurity.js')(Polyglot);
   const VehicleClimate = require('./VehicleClimate.js')(Polyglot);
+  const VehicleConditioning = require('./VehicleConditioning.js')(Polyglot);
 
   class Controller extends Polyglot.Node {
     // polyInterface: handle to the interface
@@ -101,6 +102,8 @@ module.exports = function(Polyglot) {
       logger.debug('ControllerNodes.updateOtherNodes(%s)pushdata', vehicleNodeAddress);
       const climateNode = this.polyInterface.getNode("c" + vehicleNodeAddress);
       climateNode.pushedData(vehicleId, vehicleMessage);
+      const condNode = this.polyInterface.getNode("ac" + vehicleNodeAddress);
+      condNode.pushedData(vehicleId, vehicleMessage);
     }
 
     // pass the Tesla API vehicle object
@@ -125,7 +128,6 @@ module.exports = function(Polyglot) {
               id); // We save the ID in GV20 for eventual API calls
 
           await newVehicle.initializeUOM();
-
           const result = await this.polyInterface.addNode(newVehicle);
 
           logger.info('Vehicle added: %s', result);
@@ -147,7 +149,6 @@ module.exports = function(Polyglot) {
                 id); // We save the ID in GV20 for eventual API calls
 
             const resultSecurity = await this.polyInterface.addNode(newVehicleSecurity);
-            await newVehicleSecurity.query(true); // get current values
 
             logger.info('VehicleSecurity added: %s', resultSecurity);
             this.polyInterface.addNoticeTemp(
@@ -167,9 +168,8 @@ module.exports = function(Polyglot) {
                   vehicleClimateName,
                   id); // We save the ID in GV20 for eventual API calls
 
-              const resultClimate = await this.polyInterface.addNode(newVehicleClimate);
               await newVehicleClimate.initializeUOM();
-              await newVehicleClimate.query(true); // get current values
+              const resultClimate = await this.polyInterface.addNode(newVehicleClimate);
 
               logger.info('VehicleClimate added: %s', resultClimate);
               this.polyInterface.addNoticeTemp(
@@ -177,6 +177,28 @@ module.exports = function(Polyglot) {
                 'New node created: ' + vehicleClimateName,
                 5
               );
+              
+              const vehicleConditioningName = vehicle.display_name + " Auto Conditioning";
+              const vehicleConditioningAddress =  "ac" + deviceAddress;
+              logger.info('Adding VehicleConditioning node %s: %s',
+                  vehicleConditioningAddress, vehicleConditioningName);
+                const newVehicleConditioning = new VehicleConditioning(
+                    this.polyInterface,
+                    this.address, // primary
+                    vehicleConditioningAddress,
+                    vehicleConditioningName,
+                    id); // We save the ID in GV20 for eventual API calls
+
+                const resultConditioning = await this.polyInterface.addNode(newVehicleConditioning);
+
+                logger.info('VehicleConditioning added: %s', resultConditioning);
+                this.polyInterface.addNoticeTemp(
+                  'newVehicleConditioning-' + vehicleConditioningAddress,
+                  'New node created: ' + vehicleConditioningName,
+                  5
+                );
+
+              await newVehicle.queryNow(); // get current values - will update all nodes
 
           return { added: true };