+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+V2.2.0 (2020-02-09)
+* Limit the amount of time the system may be in the Wake mode to the long poll time
+* Fixed optional rear seat heater handling
+* Added individual security overrides for lock, sunroof, charge\_port, windows, trunk, frunk, sentry, and software\_update
+* Renamed Vehicle Online to Wake Mode - true when the app is using the short poll to get updates
+* Added Vehicle State which tells whether the vehicle is actually online or asleep
+* Removed duplicate Vehicle Online from Climate and Security nodes
+* Better reporting of Charging State (Stopped, Disconnected, Charging)
+* Added Sunroof unavailable status
+* Added Conditioning node for easier command control of the auto conditioning feature
+
 v2.1.0 (2020-01-18)
 * When the General node is in wake mode, all three nodes will be refreshed including the Climate and Security nodes
 * Added configuration parameter "Enable Security Commands"